Use repository.git_http_url instead of repository.url to identify a repository.
The repository.url property in the GitLab push body specifies the SSH URL of the repository. If in Jenkins, the repository is configured to pull from HTTP, it won't match the push request as being from the checkout. By using the repository.git_http_url property, it is able to match it correctly.